chore: integrate separated endpoint with FE
